+++ src/number/downto/define.js
@@ -1,0 +1,3 @@
+import { defineInstance } from '../../core/number';
+import downto from './';
+defineInstance('downto', downto);