+++ src/number/upto/define.js
@@ -1,0 +1,3 @@
+import { defineInstance } from '../../core/number';
+import upto from './';
+defineInstance('upto', upto);