+++ src/number/util/clamp.js
@@ -1,14 +1,21 @@
 import { isNumber } from '../../util/typeChecks';
-import { assertValidNumber } from '../../util/assertions';
+import { assertNumber } from '../../util/assertions';
 
-export default function(n, min, max) {
-  assertValidNumber(n);
-  if (!isNumber(min)) {
-    min = n;
+export default function(n, ...args) {
+  const [min, max] = collectArgs(args);
+  assertNumber(n);
+  if (isNumber(min)) {
+    n = Math.max(n, min);
   }
-  if (!isNumber(max)) {
-    max = min;
-    min = n;
+  if (isNumber(max)) {
+    n = Math.min(n, max);
   }
-  return Math.min(max, Math.max(min, n));
+  return n;
+}
+
+function collectArgs(args) {
+  if (args.length === 2) {
+    return [args[0], args[1]];
+  }
+  return [null, args[0]];
 }