+++ src/util/assertions.js
@@ -10,40 +10,32 @@ function usesScientificNotation(n) {
   return abs >= MAX || abs <= MIN;
 }
 
-function isFinite(n) {
-  return Number.isFinite(n);
-}
-
-function isInteger(n) {
-  return Number.isInteger(n);
-}
-
-export function assertValidNumber(n) {
+export function assertNumber(n) {
   if (!n && n !== 0) {
     throw new TypeError(n + ' is not a valid number');
   }
 }
 
 export function assertFinite(n) {
-  if (!isFinite(n)) {
+  if (!Number.isFinite(n)) {
     throw new TypeError(n + ' is not a finite number');
   }
 }
 
 export function assertInteger(n) {
-  if (!isInteger(n)) {
+  if (!Number.isInteger(n)) {
     throw new TypeError(n + ' is not an integer');
   }
 }
 
 export function assertPositiveInteger(n) {
-  if (!isInteger(n) || n <= 0) {
+  if (!Number.isInteger(n) || n <= 0) {
     throw new TypeError(n + ' is not a positive integer');
   }
 }
 
 export function assertDecimal(n) {
-  if (!isFinite(n) || usesScientificNotation(n)) {
+  if (!Number.isFinite(n) || usesScientificNotation(n)) {
     throw new TypeError(n + ' cannot be represented as a decimal');
   }
 }